Hide compression results reference when not executing a comrpession benchmark (#17)
Removal of ramdisk and fix documentation
Use /dev/shm
Make reset and install optional
This enables us not to reset after each run and not to install on repeated runs.
Default command for install, test and reset would be: comp-test.sh -i -r

+++ comp-test.sh
@@ -21,6 +21,8 @@ fi
 now=$(date +%s)
 
 MODE="NONE"
+INSTALL="FALSE"
+RESET="FALSE"
 GZIP="gzip gzip-1 gzip-2 gzip-3 gzip-4 gzip-5 gzip-6 gzip-7 gzip-8 gzip-9"
 ZSTD="zstd zstd-1 zstd-2 zstd-3 zstd-4 zstd-5 zstd-6 zstd-7 zstd-8 zstd-9 zstd-10 zstd-11 zstd-12 zstd-13 zstd-14 zstd-15 zstd-16 zstd-17 zstd-18 zstd-19"
 ZSTDFAST="zstd-fast zstd-fast-1 zstd-fast-2 zstd-fast-3 zstd-fast-4 zstd-fast-5 zstd-fast-6 zstd-fast-7 zstd-fast-8 zstd-fast-9 zstd-fast-10 zstd-fast-20 zstd-fast-30 zstd-fast-40 zstd-fast-50 zstd-fast-60 zstd-fast-70 zstd-fast-80 zstd-fast-90 zstd-fast-100 zstd-fast-500 zstd-fast-1000"
@@ -59,11 +61,11 @@ while getopts "p:t:ribfhc:" OPTION; do
 			esac
 			;;
                 r)
-                        MODE="RESET"
+                        RESET="TRUE"
                         echo "Selected RESET of ZSTD test-installation"
                         ;;
                 i)
-                        MODE="INSTALL"
+                        INSTALL="TRUE"
                         echo "Selected INSTALL of ZSTD test-installation"
                         ;;
                 b)
@@ -118,22 +120,17 @@ done
 echo "checking if you git cloned zfs"
 [ ! -e ./zfs/.git ] && { echo "You need to clone zfs first! # git clone https://github.com/zfsonlinux/zfs"; exit 1; }
 
-if [ $MODE = "RESET" ]
+
+
+if [ $INSTALL = "TRUE" ]
 then
-        echo "destroy testpool and unmount ramdisk of previous broken/canceled tests"
-        test -f ./zfs/cmd/zpool/zpool && sudo ./zfs/cmd/zpool/zpool destroy testpool 2>&1 >/dev/null
-        sudo umount -l /mnt/ramdisk >/dev/null 2>&1
 
         cd ./zfs
-        echo "make sure zfs is unloaded and make is cleaned"
+        echo "unloading and unlinking possible previous build"
         sudo ./scripts/zfs.sh -u
         sudo ./scripts/zfs-helpers.sh -r
         make -s distclean >> /dev/null
-fi
-
-if [ $MODE = "INSTALL" ]
-then
-        cd ./zfs
+        
         echo "rebuilding zfs"
         sh autogen.sh >> /dev/null
         ./configure --enable-debug >> /dev/null
@@ -142,40 +139,20 @@ then
         echo "loading zfs"
         sudo ./scripts/zfs-helpers.sh -i
         sudo ./scripts/zfs.sh
+        cd ..
 fi
 
 if [  $MODE = "FULL" -o $MODE = "BASIC" -o $MODE = "CUSTOM" ]
 then
-        echo "destroy testpool and unmount ramdisk of previous broken/canceled tests"
+        echo "destroy testpool and cleam ram of previous broken/canceled tests"
         test -f ./zfs/cmd/zpool/zpool && sudo ./zfs/cmd/zpool/zpool destroy testpool 2>&1 >/dev/null
-        sudo umount -l /mnt/ramdisk >/dev/null 2>&1
-
-        cd ./zfs
-        echo "make sure zfs is unloaded and make is cleaned"
-        sudo ./scripts/zfs.sh -u
-        sudo ./scripts/zfs-helpers.sh -r
-        make -s distclean >> /dev/null
-
-
-        echo "rebuilding zfs"
-        sh autogen.sh >> /dev/null
-        ./configure --enable-debug >> /dev/null
-        make -s -j$(nproc) >> /dev/null
-
-        echo "loading zfs"
-        sudo ./scripts/zfs-helpers.sh -i
-        sudo ./scripts/zfs.sh
-
-        cd ..
-        echo "creating ramdisk"
-        sudo mkdir /mnt/ramdisk
-        sudo mount -t tmpfs -o size=2400m tmpfs /mnt/ramdisk
+        sudo rm -f /dev/shm/pooldisk.img
 
         echo "creating virtual pool drive"
-        truncate -s 1200m /mnt/ramdisk/pooldisk.img
+        truncate -s 1200m /dev/shm/pooldisk.img
 
         echo "creating zfs testpool/fs1"
-        sudo ./zfs/cmd/zpool/zpool create testpool -f -o ashift=12  /mnt/ramdisk/pooldisk.img
+        sudo ./zfs/cmd/zpool/zpool create testpool -f -o ashift=12  /dev/shm/pooldisk.img
         sudo ./zfs/cmd/zfs/zfs create testpool/fs1
 		sudo ./zfs/cmd/zfs/zfs set recordsize=1M  testpool/fs1
 
@@ -201,16 +178,16 @@ then
 			;;
 	esac
 
-        echo "copying $FILENAME to ramdisk, truncating it after 1000M"
-	sudo dd if=$FILENAME of=/mnt/ramdisk/$FILENAME bs=1M count=1000 status=none
-        cd /mnt/ramdisk/
+        echo "copying $FILENAME to ram, truncating it after 1000M"
+	    sudo dd if=$FILENAME of=/dev/shm/$FILENAME bs=1M count=1000 status=none
+        cd /dev/shm/
         chksum=`sha256sum $FILENAME`
         cd -
         echo "" >> "./$TESTRESULTS"
         echo "Test with $FILENAME file" >> "./$TESTRESULTS"
-	grep "^model name" /proc/cpuinfo |sort -u >> "./$TESTRESULTS"
-	grep "^flags" /proc/cpuinfo |sort -u >>  "./$TESTRESULTS"
-	echo "" >> "./$TESTRESULTS"
+	    grep "^model name" /proc/cpuinfo |sort -u >> "./$TESTRESULTS"
+	    grep "^flags" /proc/cpuinfo |sort -u >>  "./$TESTRESULTS"
+	    echo "" >> "./$TESTRESULTS"
 
         echo "starting compression test suite"
         for comp in $ALGO
@@ -222,7 +199,7 @@ then
 			echo "Could not set compression to $comp! Skipping test."
 		else
                 	echo “Compression results for $comp” >> "./$TESTRESULTS"
-                	dd if=/mnt/ramdisk/$FILENAME of=/testpool/fs1/$FILENAME bs=4M 2>&1 |grep -v records >> "./$TESTRESULTS"
+                	dd if=/dev/shm/$FILENAME of=/testpool/fs1/$FILENAME bs=4M 2>&1 |grep -v records >> "./$TESTRESULTS"
                 	./zfs/cmd/zfs/zfs get compressratio testpool/fs1 >> "./$TESTRESULTS"
                 	echo "" >> "./$TESTRESULTS"
                 	echo “Decompression results for $comp” >> "./$TESTRESULTS"
@@ -241,15 +218,25 @@ then
         done
 
         echo "compression test finished"
-        echo "destroying pool and unmounting ramdisk"
-        sudo ./zfs/cmd/zpool/zpool destroy testpool
-        sudo umount -l /mnt/ramdisk
+        echo "destroying pool"
+        test -f ./zfs/cmd/zpool/zpool && sudo ./zfs/cmd/zpool/zpool destroy testpool 2>&1 >/dev/null
+        echo "Cleaning ram"
+        sudo rm -f /dev/shm/pooldisk.img
+
+fi
+
+if [ $RESET = "TRUE" ]
+then
+        cd ./zfs
         echo "unloading and unlinking zfs"
-        cd zfs
         sudo ./scripts/zfs.sh -u
         sudo ./scripts/zfs-helpers.sh -r
-        make distclean >> /dev/null
+        make -s distclean >> /dev/null
         cd ..
+fi
+echo "Done."
 
-        echo "Done. results written to ./$TESTRESULTS"
+if [  $MODE = "FULL" -o $MODE = "BASIC" -o $MODE = "CUSTOM" ]
+then
+echo "compression results written to ./$TESTRESULTS"
 fi
